DDBTEAM-1040: Refactored vendor interfaces to allow vendors with only "single identifier" support
