+++ config/services.yaml
@@ -36,8 +36,11 @@ services:
             calls:
                 - setVendorCoreService: ['@App\Service\VendorService\VendorCoreService']
 
+        App\Service\VendorService\VendorServiceImporterInterface:
+            tags: [app.vendor.service.importer]
+
         App\Service\VendorService\VendorServiceSingleIdentifierInterface:
-            tags: [ app.handler.single_identifier_vendor ]
+            tags: [ app.vendor.service.single_identifier ]
 
     # makes classes in src/ available to be used as services
     # this creates a service per class whose id is the fully-qualified class name
@@ -95,4 +98,4 @@ services:
 
     App\MessageHandler\SearchNoHitsMessageHandler:
         arguments:
-            $singleIdentifierVendors: !tagged_iterator app.handler.single_identifier_vendor
+            $singleIdentifierVendors: !tagged_iterator app.vendor.service.single_identifier
